Also move User CLI tests into suite.
