+++ src/userCLI/src/test/kotlin/org/vaccineimpact/api/security/tests/GroupsTests.kt
@@ -1,10 +1,11 @@
-package org.vaccineimpact.api.security
+package org.vaccineimpact.api.security.tests
 
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import org.vaccineimpact.api.db.JooqContext
 import org.vaccineimpact.api.db.direct.addGroup
+import org.vaccineimpact.api.security.Groups
 import org.vaccineimpact.api.test_helpers.DatabaseTest
 
 class GroupsTests : DatabaseTest()
